Add files via upload
This software has two parts. First is user part and the 
administrator part. User part is used as a front end and 
administrator is the back end. Administrator is used by airline 
authority. It will allow the customers to access database and allow 
new customers to sign up for online access.  
The system allows the airline passenger to search for flights that 
are available between the two travel cities, namely the “Departure 
city” and “Arrival city” for a particular departure and arrival 
dates. The system displays all the flight’s details such as flight no, 
name, price and duration of journey etc. After search the system 
display list of available flights and allows customer to choose a 
particular flight. Then the system checks for the availability of 
seats on the flight. If the seats are available then the system 
allows the passenger to book a seat. Otherwise, it asks the user to 
choose another flight. To book a flight the system asks the 
customer to enter his details such as name, address, city, state, 
and credit card number and contact number. Then it checks the 
validity of card and book the flight and update the airline 
database and user database. The system also allows the customer 
to cancel his/her reservation, if any problem occurs.  
The main purpose of this software is to reduce the manual errors 
involved in the airline reservation process and make it convenient 
for the customers to book the flights as when they require such 
that they can utilize this software to make reservations, modify 
reservations or cancel a particular reservation.